Ensure persistence.driver option exists before validating
The driver option is not required by the configuration, so we should check for it the validation functions.

If type_prototype is being used, there is a legitimate case where type definitions would not specify a driver. Likewise, the type_prototype may not specify the driver. If we wish to validate that a driver has been specified for a type, we'll have to do that in the extension class after loading the configuration and merging in prototypes.