Ensure Doctrine objects with collection changes are updated
Added support to the Doctrine Listener to ensure that when a Collection (e.g. ReferenceMany) is modified on an Entity/Document it is updated in Elasticsearch. Moved Entity/Document scheduling logic for update and insertion into their own methods and keyed the schedule arrays by the spl_object_hash of the object.