Fix pagination of ORM queries when populating
Paginating a DQL query which does not have a ORDER BY clause does not
work cross-platform, as the sorting is undefined in such case.
MySQL is not affected by the bug because it always adds an implicit sorting
by insertion order (which is a compliant way to implement the undefined order
but tends to hide this issue from developers having to deal with other
platforms too).