Address deprecations from persistence (#1600)
* Address deprecations from persistence

A backwards-compatibility layer has been added to persistence to help
consumers move to the new namespacing. It is based on class aliases,
which means the type declaration changes should not be a BC-break: types
are the same.
See doctrine/persistence#71

This means:
- using the new namespaces
- adding autoload calls for new types to types that may be extended and
use persistence types in type declarations of non-constructor methods,
so that signature compatibility is recognized by old versions of php.
More details on this at
https://dev.to/greg0ire/how-to-deprecate-a-type-in-php-48cf

* Remove trailing whitespace

Conflicts:
	composer.json
	tests/Unit/Doctrine/PHPCR/ElasticaToModelTransformerTest.php