Restored pagerfanta/pagerfanta dependency (#1912)
* Moved pagerfanta adapters to "required" section of the composer.json

Since adapters are needed for the functionality they should be installed automatically otherwise exceptions are thrown. 

Feel free to close the PR if that was intentional and we have to specify the adapters we need from now on manually. This should be mentioned in the documentation and in the CHANGELOG though.

* Required pagerfanta full-package instead of only core